pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "django-vendor"
version = "0.5.0dev1"
description = "Django App Toolkit for selling digital and physical goods online."
authors = [
    { name = "Grant Viklund", email = "renderbox@gmail.com" },
    { name = "Roberto Himmelbauer" }
]
readme = "README.md"
requires-python = ">=3.10"
keywords = ["django", "app"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12"
]
dependencies = [
    "django>=5.2,<7.0",
    "iso4217<1.12.0"
]

[project.optional-dependencies]
stripe = ["stripe"]
authorizenet = ["authorizenet"]
test = [
    "pytest",
    "pytest-django",
    "pytest-cov",
    "pytest-factoryboy",
    "pytest-mock",
    "pytest-asyncio",
    "pytest-bdd",
    "django-vendor[stripe]",
    "django-vendor[authorizenet]"
]
lint = [
    "black",
    "flake8",
    "flake8-black",
    "isort",
    "mypy",
    "bandit"
]
dev = [
    "django-vendor[test]",
    "django-vendor[lint]",
    "django-crispy-forms",
    "django-site-configs",
    "django-integrations",
    "django-allauth"
]

[project.urls]
"Homepage" = "https://github.com/renderbox/django-vendor"
"Bug Tracker" = "https://github.com/renderbox/django-vendor/issues"
"Documentation" = "https://github.com/renderbox/django-vendor"

[tool.setuptools.packages.find]
where = ["src"]
include = ["vendor*"]

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
minversion = "6.0"
